moved error-page-container in contaienrs, resolve style issues
