+++ packages/fe-container/src/components/error-page-500/ErrorPage500.tsx
@@ -15,7 +15,7 @@
  */
 import React from 'react'
 
-import {ErrorPageContainer} from '@components/error-page-container/ErrorPageContainer'
+import {ErrorPageContainer} from '../../containers/error-page/ErrorPageContainer'
 import {ReactComponent as Error500Logo} from './assets/ErrorImage500.svg'
 
 export const ErrorPage500: React.FC = () => {
