+++ packages/fe-container/src/containers/error-page/ErrorPageContainer.less
@@ -1,4 +1,5 @@
 @import "../../styles/_variables";
+@import "../../styles/_colors.less";
 
 .errorPage_container {
     height: 100%;
@@ -9,23 +10,21 @@
     flex-direction: column;
     
     & .mainMessage {
-        font-family: 'Proxima Nova';
         font-style: normal;
         font-weight: 600;
         font-size: 24px;
         line-height: 34px;
         text-align: center;
-        color: var(--error-page-text-color);
+        color: @error-page-text-color
     }
 
     & .descriptionMessage {
-        font-family: 'Proxima Nova';
         font-style: normal;
         font-weight: normal;
         font-size: 20px;
         line-height: 28px;
         text-align: center;
-        color: var(--error-page-text-color);
+        color: @error-page-text-color;
     }
 }
