+++ packages/fe-container/src/styles/_colors.less
@@ -4,9 +4,9 @@
 @menu-label-color: #666666;
 @menu-hover-bg-color: rgba(45, 168, 155, 0.2);
 @menu-user-logout-color: #f54b42;
+@error-page-text-color : #0E0E0E;
 
 :root {
   --microlc-primary-color: #1890FF;
   --microlc-menu-entry-selected-bg-color: #E6F3FF;
-  --error-page-text-color : #0E0E0E;
 }
